Output tibble after applying all IND~pressure modeling functions

This is an example output tibble based on the Central Baltic Sea food web indicator demonstration data after applying the calc_deriv and the test_interaction functions on the merge_models_ex tibble.

all_results_ex

Format

A data frame with 84 rows and 31 variables:

id

Numerical IDs for the IND~press combinations.

ind

Indicator names.

press

Pressure names.

model_type

Specification of the model type; at this stage containing only "gam" (Generalized Additive Model).

corrstruc

Specification of the correlation structure; at this stage containing only "none".

aic

AIC of the fitted models

edf

Estimated degrees of freedom for the model terms.

p_val

The p values for the smoothing term (the pressure).

signif_code

The significance codes for the p-values.

r_sq

The adjusted r-squared for the models. Defined as the proportion of variance explained, where original variance and residual variance are both estimated using unbiased estimators. This quantity can be negative if your model is worse than a one parameter constant model, and can be higher for the smaller of two nested models.

expl_dev

The proportion of the null deviance explained by the models.

nrmse

Absolute values of the root mean square error normalized by the standard deviation (NRMSE) and corrected for the prior (log) transformation.

ks_test

The p-values from a Kolmogorov-Smirnov Test applied on the model residuals to test for normal distribution. P-values > 0.05 indicate normally distributed residuals.

tac

logical; indicates whether temporal autocorrelation (TAC) was detected in the residuals. TRUE if model residuals show TAC.

pres_outlier

A list-column with outliers identified for each model (i.e. Cook`s distance > 1). The indices present the position in the training data, including NAs.

excl_outlier

A list-column listing all outliers per model that have been excluded in the GAM fitting

model

A list-column of IND~press-specific gam objects.

prop

The proportion of the observed pressure range where the indicator shows a response (see the last section in Details)

zero_in_conf

A list-column of logical vectors indicating for every pressure value (in press_seq) whether the slope of the indicator response at that pressure value is within the confidence interval, i.e. is zero.

zic_start_end

A list-column of logical vectors indicating for every pressure value (in press_seq) whether the slope is considered as zero for the proportion calculation (see see the last section in Details)

.
press_seq

A list-column with sequences of 100 evenly spaced pressure values.

pred

A list-column with the predicted indicator responses averaged across all bootstraps (for the 100 equally spaced pressure values).

pred_ci_up

A list-column with the upper confidence limit of the bootstrapped predictions.

pred_ci_low

A list-column with the lower confidence limit of the bootstrapped predictions.

deriv1

A list-column with the first derivatives of the indicator responses averaged across all bootstraps (for the 100 equally spaced pressure values).

deriv1_ci_up

A list-column with the upper confidence limit of the bootstrapped first derivatives.

deriv1_ci_low

A list-column with the lower confidence limit of the bootstrapped first derivatives.

adj_n_boot

The number of successful bootstrap samples that was actually used for calculating the mean and confidence intervals of the predicted indicator response and the derivative.

boot_error

A list-column capturing potential error messages that occurred as side effects when refitting the GAM(M)s on each bootstrap sample.

interaction

logical; if TRUE, at least one thresh_gam performs better than its corresponding gam based on the leave-one-out cross-validation.

thresh_var

A list-column with the threshold variables of the better performing thresh_models.

thresh_models

A list-column with nested lists containing the better performing thresh_models.

thresh_error

A list-column capturing potential error messages that occurred as side effects when fitting each threshold GAMs and performing the LOOCV.

tac_in_thresh

logical vector; indicates for every listed thresh_model whether temporal autocorrelation (TAC) was detected in the residuals. TRUE if model residuals show TAC.