Check class and variable names and types of ind and press data

check_ind_press is a helper function for model_trend and ind_init. It coerces the data into a data frame and checks the type of each variable. If any of the variables is not a number (double or integer) it will return an error message. Also the variable names are checked for unwanted characters (e.g. hyphen, brackets, etc.) and when necessary replaced by an underscore. The latter is required for the correct model specifications.

check_ind_press(x, input = "ind")

Arguments

x
A data object containing only the indicator or pressure variables (vector, matrix, data frame or tibble).

input
Indication whether x represents the indicator (default) or pressure variable(s). If x is a vector, `input` determines the variable name when coerced to a data frame.

Value

The function returns the input object as data frame with correct variable names.

See also