Score-based cluster analysis

clust_sc computes a hierarchical cluster analysis for the identification of indicator redundancies.

clust_sc(dist_mat, method_clust = "average", ...)

Arguments

dist_mat
The distance matrix computed by the dist_sc function.

method_clust
The agglomeration method to be used in the hclust function. Default is "average", for alternatives see hclust.

...
Further arguments to be passed to the method hclust.

Value

An object of class hclust is returned, which describes the tree produced by the clustering process. See for more details hclust. Additionally, the cophenetic correlation coefficient and the Gower distance are printed in the console as guidance for selecting the best agglomeration method.

See also

Examples

# Using the Baltic Sea demo data
scores_tbl <- scoring(trend_tbl = model_trend_ex,
  mod_tbl = all_results_ex, press_type = press_type_ex)
scores_mat <- summary_sc(scores_tbl)$scores_matrix
dist_matrix <- dist_sc(scores_mat, method_dist = "euclidean")
clust_analysis <- clust_sc(dist_matrix, method_clust = "complete")

#> With the selected agglomeration method (complete), the cophonetic correlation coeffient is 0.842 and the Gower distance is 417.41.