Template of (sub-)criteria and corresponding scoring information

This table serves as basis for the scoring function and builds on the criterion-scoring scheme described in the underlying framework (Otto et al., 2018). The user can modify the weights, scores, conditions or remove specific (sub-)crits.

crit_scores_tmpl

Format

A data frame with 27 rows and 12 variables:

crit_id

ID of main criteria.

crit

A vector of the main criteria.

subcrit_id

ID of sub-criteria.

subcrit

A vector of the sub-criteria.

definition

A short description of the (sub)criteria.

score_explanation

A short explanation on the required condition for a specific score.

score

A vector of scores for each (sub-)criterion and the respective condition.

weight

Weights assigned to each score to allow easy user adjustments. Default is 1 for all scores.

score_pressure_specific

Additional information whether the scoring is pressure-specific or not.

condition

A list-column single elements or vectors with various elements containing score-specific conditions put in R Syntax against which (sub-)criterion-specific variables (see condition_var) are checked.

condition_var

The variables used as basis for scoring the specific (sub-)criterion.

func_name

The names of the function that generate output tibbles containing the required variables.