Score-based distance matrix

The function computes a distance matrix based on the scoring output tibble (or the output tibble from the expect_resp function).

dist_sc(scores_mat, scores_tbl, method_dist = "euclidean", ...)

Arguments

scores_mat
A data frame or matrix containing the scores. This could be the $scores_matrix output of the summary_sc function.

scores_tbl
Deprecated argument from earlier version (0.1.0); scores_tbl represented the output tibble from the scoring function.

method_dist
Dissimilarity index used in the vegdist function to calculate the dissimilarity matrix based on the scores. Default is `euclidean`, for alternatives see vegdist.

...
Further arguments to be passed to the method vegdist.

Value

The function returns a dist object.

See also

Examples

# Using the Baltic Sea demo data
scores_tbl <- scoring(trend_tbl = model_trend_ex,
  mod_tbl = all_results_ex, press_type = press_type_ex)
scores_mat <- summary_sc(scores_tbl)$scores_matrix
dist_matrix <- dist_sc(scores_mat, method_dist = "euclidean")