Score adjustments for sub-criterion 10.1

expect_resp runs a shiny app in which the expectation of the IND response to a pressure (sub-criterion 10.1) can be manually changed to `yes` or `no` based on visual inspection of the IND response curve.

expect_resp(mod_tbl, scores_tbl, crit_scores = INDperform::crit_scores_tmpl)

Arguments

mod_tbl
Output tibble from the IND~pressure modeling functions.

scores_tbl
The output tibble from the scoring function.

crit_scores
The(un)modified criterion-scoring template crit_scores_tmpl; has to be the same than used in scoring. Default is the unmodified template crit_scores_tmpl.

Value

The function returns the input scoring tibble, but with modified scores in the variable C10_1, once the "Press Me!" button is activated.

Details

The sub-criterion 10.1 (i.e. the IND response to a pressure, which has been found significant, is in line with expectations based on ecological knowledge) has been set to a default score of 1 (no expectation / unclear as response is highly non-linear) in the scoring function. Determining whether the IND response modeled in the GAM/GAMM meets specific expectations can only be done based on visual model inspections. expect_resp provides only a very simple graphical representation of this smoothing function.

For a more comprehensive figure use the plot_model function and then go back to this function for modifications of the expectation scores.

See also

Examples

if (FALSE) {
 # Using the Baltic Sea demo data:
 # Apply first the scoring on the model outputs
 scores_tbl <- scoring(trend_tbl = model_trend_ex, mod_tbl = all_results_ex,
   press_type = press_type_ex)
 # Then run the expect_resp() shiny function to correct one criterion
 scores_tbl <- expect_resp(all_results_ex, scores_tbl)
 # Check if it worked:
 expect_resp(all_results_ex, scores_tbl)
}