Output tibble from the ind_init function

This is an example output tibble from the ind_init function applied on the Central Baltic Sea food web indicator demonstration data.

ind_init_ex

Format

A data frame with 84 rows and 10 variables:

id

Numerical IDs for the IND~press combinations.

ind

Indicator names.

press

Pressure names.

ind_train

A list-column with indicator values of the training data.

press_train

A list-column with pressure values of the training data.

time_train

train data from year

ind_test

A list-column with indicator values of the test data.

press_test

A list-column with pressure values of the test data.

time_test

test data from year

train_na

logical; indicates the joint missing values in the training IND and pressure data. That includes the original NAs as well as randomly selected test observations that are within the training period