Merging two model output tibbles.

The function appends the second model output tibble to the first while keeping all variables from both tibbles.

merge_models(mod_tbl1, mod_tbl2)

Arguments

mod_tbl1
Model output tibble created e.g. with model_gam.

mod_tbl2
Model output tibble created e.g. with select_model.

Value

merge_models returns the same type as the input including all columns of both tibbles.

Details

merge_models function applies internally the dplyr::bind_rows function so that columns are matched by name, and any missing columns will be filled with NA. The function has also some data validation incorporated to check for double entries.

Examples

# Using some models of the Baltic Sea demo data:
# Merging GAM and GAMM tibbles
test_ids <- 47:50 # choose subset
gam_tbl <- model_gam_ex[test_ids,]
gamm_tbl <- model_gamm(ind_init_ex[test_ids,], filter= gam_tbl$tac)

#> NOTE: For the following IND~pressure GAMMs fitting procedure failed:

#> # A tibble: 1 x 6
#>      id ind   press model_type corrstruc error_message                          
#>   <int> <chr> <chr> <chr>      <chr>     <chr>                                  
#> 1    48 Sprat Fher  gamm       arma21    "Error in lme.formula(y ~ X - 1, rando…

best_gamm <- select_model(gam_tbl, gamm_tbl)
merge_models(gam_tbl[gam_tbl$tac == FALSE,], best_gamm)

#> # A tibble: 4 x 17
#>      id ind   press model_type corrstruc   aic   edf   p_val signif_code    r_sq
#>   <int> <chr> <chr> <chr>      <chr>     <dbl> <dbl>   <dbl> <chr>         <dbl>
#> 1    47 Sprat Fspr… gamm       ar1        66.5 1.14  8.13e-1 " "         -0.0702
#> 2    48 Sprat Fher  gamm       ar2        60.4 3.85  6.66e-5 ***          0.364 
#> 3    49 Sprat Fcod  gamm       ar1        63.6 1.000 9.27e-2 .           -0.0170
#> 4    50 Herr… Tsum  gam        none       24.6 1.    2.49e-1 " "          0.0148
#> # … with 7 more variables: expl_dev <dbl>, nrmse <dbl>, ks_test <dbl>,
#> #   tac <lgl>, pres_outlier <list>, excl_outlier <list>, model <list>

# Merge 2 IND-specific GAM tibbles (where)
dat_init <- ind_init(
  ind_tbl = ind_ex[, c("TZA", "Cod")],
  press_tbl = press_ex[, c("Tsum", "Swin")],
  time = ind_ex[,1])
gam_tbl1 <- model_gam(dat_init[1:2, ])
# treat a subset differently, e.g. when setting k
gam_tbl2 <- model_gam(dat_init[3:4, ], k = 3)
merge_models(gam_tbl1, gam_tbl2)

#> # A tibble: 4 x 16
#>      id ind   press model_type corrstruc   aic   edf  p_val signif_code    r_sq
#>   <int> <chr> <chr> <chr>      <chr>     <dbl> <dbl>  <dbl> <chr>         <dbl>
#> 1     1 TZA   Tsum  gam        none       49.8  1.92 0.146  " "          0.138 
#> 2     2 TZA   Swin  gam        none       48.1  2.74 0.0772 .            0.211 
#> 3     3 Cod   Tsum  gam        none       68.4  1.   0.648  " "         -0.0312
#> 4     4 Cod   Swin  gam        none       64.1  1.57 0.102  " "          0.139 
#> # … with 6 more variables: expl_dev <dbl>, nrmse <dbl>, ks_test <dbl>,
#> #   tac <lgl>, pres_outlier <list>, model <list>