Summary of indicator performance scores

Summarizes the scoring output tibble so that IND-specific scores for each criterion as well as the pressure-specific sub-criteria scores (in crit. 9 and 10) can be easily compared and used for further score-based IND performance functions.

summary_sc(scores_tbl, crit_scores = INDperform::crit_scores_tmpl)

Arguments

scores_tbl
The output tibble from the scoring function.

crit_scores
The (un)modified criterion-scoring template crit_scores_tmpl; required to calculate the scores in percentage. Has to be the same than used in scoring. Default is the unmodified template crit_scores_tmpl.

Value

The function returns a list of 3 data frames

overview

IND-specific scores and percentages from max. score for all criteria (crit 9 and 10 averaged across all sign. pressures and the number of significant pressures).

subcriteria_per_press

IND- and pressure-specific scores for all (sub-)criteria and the percentages from max. criterion score.

scores_matrix

TEXT

See also

Examples

# Using the Baltic Sea demo data in this package
scores_tbl <- scoring(trend_tbl = model_trend_ex, mod_tbl = all_results_ex,
  press_type = press_type_ex)
summary_sc(scores_tbl)

#> $overview
#>        ind nr_sign_press C8  C9 C10 C11 C8_in% C9_in% C10_in% C11_in%
#> 1      TZA             2  0 5.0 2.0   2      0     83      33     100
#> 2       MS             2  0 4.0 2.5   1      0     67      42      50
#> 3      rCC             1  0 6.0 4.0   2      0    100      67     100
#> 4     Cops             0  0 0.0 0.0   0      0      0       0       0
#> 5    Micro             2  1 5.0 4.0   1    100     83      67      50
#> 6    rZPPP             1  0 4.0 5.0   2      0     67      83     100
#> 7    Sprat             1  1 2.0 1.0   2    100     33      17     100
#> 8  Herring             0  1 0.0 0.0   0    100      0       0       0
#> 9  Stickle             0  1 0.0 0.0   0    100      0       0       0
#> 10     Cod             2  1 5.5 1.5   2    100     92      25     100
#> 11     SPF             1  1 5.0 4.0   2    100     83      67     100
#> 12     LPF             1  1 5.0 5.0   2    100     83      83     100
#> 
#> $subcriteria_per_press
#>      ind  press     press_type C9_1 C9_2 C10_1 C10_2 C10_3 C10_4 C9 C10 C9_in%
#> 1    TZA Fsprat        Fishing    3    3     1     0     0     0  6   1    100
#> 2    TZA   Fher        Fishing    1    3     1     1     1     0  4   3     67
#> 3     MS   Swin        Climate    2    2     1     0     1     1  4   3     67
#> 4     MS   Fher        Fishing    1    3     1     1     0     0  4   2     67
#> 5    rCC   Tsum        Climate    3    3     1     1     1     1  6   4    100
#> 6  Micro   Tsum        Climate    2    3     1     1     2     1  5   5     83
#> 7  Micro Fsprat        Fishing    2    3     1     0     2     0  5   3     83
#> 8  rZPPP   Tsum        Climate    1    3     1     1     2     1  4   5     67
#> 9  Sprat   Fher        Fishing    2    0     1     0     0     0  2   1     33
#> 10   Cod   Fher        Fishing    3    3     1     0     0     0  6   1    100
#> 11   Cod   Fcod        Fishing    2    3     1     0     1     0  5   2     83
#> 12   SPF   Pwin Eutrophication    2    3     1     1     1     1  5   4     83
#> 13   LPF   Fcod        Fishing    2    3     1     1     2     1  5   5     83
#>    C10_in%
#> 1       17
#> 2       50
#> 3       50
#> 4       33
#> 5       67
#> 6       83
#> 7       50
#> 8       83
#> 9       17
#> 10      17
#> 11      33
#> 12      67
#> 13      83
#> 
#> $scores_matrix
#>         C8 C11 Fcod_C10 Fcod_C9 Fher_C10 Fher_C9 Fsprat_C10 Fsprat_C9 Nwin_C10
#> TZA      0   2        0       0        3       4          1         6        0
#> MS       0   1        0       0        2       4          0         0        0
#> rCC      0   2        0       0        0       0          0         0        0
#> Cops     0   0        0       0        0       0          0         0        0
#> Micro    1   1        0       0        0       0          3         5        0
#> rZPPP    0   2        0       0        0       0          0         0        0
#> Sprat    1   2        0       0        1       2          0         0        0
#> Herring  1   0        0       0        0       0          0         0        0
#> Stickle  1   0        0       0        0       0          0         0        0
#> Cod      1   2        2       5        1       6          0         0        0
#> SPF      1   2        0       0        0       0          0         0        0
#> LPF      1   2        5       5        0       0          0         0        0
#>         Nwin_C9 Pwin_C10 Pwin_C9 Swin_C10 Swin_C9 Tsum_C10 Tsum_C9
#> TZA           0        0       0        0       0        0       0
#> MS            0        0       0        3       4        0       0
#> rCC           0        0       0        0       0        4       6
#> Cops          0        0       0        0       0        0       0
#> Micro         0        0       0        0       0        5       5
#> rZPPP         0        0       0        0       0        5       4
#> Sprat         0        0       0        0       0        0       0
#> Herring       0        0       0        0       0        0       0
#> Stickle       0        0       0        0       0        0       0
#> Cod           0        0       0        0       0        0       0
#> SPF           0        4       5        0       0        0       0
#> LPF           0        0       0        0       0        0       0
#>