Scoring of indicator performance

Scoring of each indicator based on an internal criterion-scoring scheme applied on the output of the trend and pressure model functions.

scoring(
  trend_tbl = NULL,
  mod_tbl,
  press_type = NULL,
  crit_scores = INDperform::crit_scores_tmpl,
  sign_level = 0.05
)

Arguments

trend_tbl
Output tibble from the model_trend function.

mod_tbl
Output tibble from the IND~pressure modeling functions.

press_type
Data frame or tibble with pressure names (named `press`) in first column and corresponding pressure types in second column (named `press_type`). Needed for the spie chart! (see for an example press_type_ex)

crit_scores
Internal tibble of (sub)criteria and respective scores named crit_scores_tmpl; can be modified by saving this data frame as new object and removing single (sub)criteria or assigning weights (default is 1). The variable `condition` represents a list of single elements or vectors with various elements to base the scoring on. This can be modified but needs to follow the same syntax.

sign_level
Significance level on which scoring is built; default is 0.05.

Value

The function returns a nested tibble with the following elements depending on the criteria rated

ind

A vector of the indicator names.

C8 and/or C11

A vector of IND-specific scores for criterion 8 (trend indication) and/or C11 (management application).

press_spec_sc

A list-column of IND-specific data frames containing pressure-specific scores for the sub-criteria 9.1-9.2, 10.1-10.4.

The tibble can easily be unnested by using the unnest function. That is, each element of the data frame in the list-column press_spec_sc becomes its own row in the tibble.

Details

Among the 16 common indicator selection criteria summarized in Otto et al. (2018) five criteria relate to the indicators` performances and require time series for their evaluation, i.e.

Crit. 8: Development reflects ecosystem change caused by variation in manageable pressure(s)

Crit. 9: Sensitive or responsive to pressures

Crit. 10: Robust, i.e. responses in a predictive fashion, and statistically sound

Crit. 11: Links to management measures (responsiveness and specificity)

Crit. 12: Relates where appropriate to other indicators but is not redundant (not scored)

In this function, the scoring scheme for these criteria as proposed by Otto et al. (2018) serves as basis for the quantification of the IND performance. Sensitivity (criterion 9) and robustness (criterion 10) are specified into more detailed sub-criteria to allow for quantification based on statistical models and rated individually for every potential pressure that might affect the IND directly or indirectly. In the case of non-significant relationships between a IND and a specific pressures, sub-crit. 9.1 and all following pressure-specific sub-crit. in criteria 9 and 10 are scored zero for this pressure.

The template tibble crit_scores_tmpl contains all relevant informations and serves as basis for the scoring in this function. See for more details crit_scores_tmpl or View(crit_scores_tmpl). The scoring scheme can easily be adapted to any kind of state indicator and management scheme by modifying the scores, the weighting of scores or by removing or adding (sub)criteria in the crit_scores_tmpl template. The condition variable can also be modified but needs to follow the same syntax.

References

Otto, S.A., Kadin, M., Casini, M., Torres, M.A., Blenckner, T. (2018) A quantitative framework for selecting and validating food web indicators. Ecological Indicators, 84: 619-631, doi: https://doi.org/10.1016/j.ecolind.2017.05.045

See also

Examples

# Using the Baltic Sea demo data in this package
scores_tbl <- scoring(trend_tbl = model_trend_ex, mod_tbl = all_results_ex,
  press_type = press_type_ex)
if (FALSE) {
 # To see the criterion template and change it potentially
 View(crit_scores_tmpl)
 # E.g. exclude the trend criterion
 crit_scores_tmpl_new <- crit_scores_tmpl[crit_scores_tmpl$crit_id > 1, ]
 # Now the trend tibble is not needed anymore
 scores_tbl <- scoring(mod_tbl = all_results_ex, press_type = press_type_ex,
  crit_scores = crit_scores_tmpl_new)
}